requirements.txt
fastapi[standard]
paho-mqtt
python-dotenv